Move back components that were moved from folder to avoid unnecessary conflicts (we'll rather do that in a different upcoming PR)
