+++ src/stories/nrn/animation/BubbleTimer.stories.tsx
@@ -3,7 +3,7 @@ import {
   Story,
 } from '@storybook/react/types-6-0';
 import React from 'react';
-import BubbleTimer, { Props } from '../../../components/animations/BubbleTimer';
+import BubbleTimer, { Props } from '../../../components/utils/BubbleTimer';
 
 export default {
   title: 'Next Right Now/Animation/BubbleTimer',
