+++ src/stories/nrn/overlay/Tooltip.stories.tsx
@@ -3,7 +3,7 @@ import {
   Story,
 } from '@storybook/react/types-6-0';
 import React from 'react';
-import Tooltip, { Props } from '../../../components/overlay/Tooltip';
+import Tooltip, { Props } from '../../../components/utils/Tooltip';
 
 export default {
   title: 'Next Right Now/Overlay/Tooltip',
